Move drag state down into redux for better readability.
This makes the code much easier to read, at the
expense of additional redux ceremony. I do not believe
that this leads to any additional storage cycles
in redux, because in general every 'active' mouse move
was already going down to the redux store.

As part of this change, VisualGraph is now a computed
value, available via a memoeised selector using
https://github.com/reduxjs/reselect