Move mid point to the middle of the shaft.
Previously it was in the middle of the whole arrow, which
was sub-optimal for arrows with large arrows heads.
This change is a compromise: originally the mid point was
set to the middle of the whole arrow so that midpoints
aligned across a whole bundle of arrows with different
directions. However, that situation is probably not
the common case, so making this change for now.