Remove graph-app-kit and upgrade react-scripts.
graph-app-kit is not really used, and in any case the global-scope-
injection mechanism has now been superseded by GraphQL. Just remove it
for now.

react-scripts has been upgraded, and this had some bizarre knock-on
effects; especially it seems that any method called 'use...' is now
considered to be a React hook. Renamed methods to work around this.