Create a CaptionEditor if no components present.
This is a better solution to the problem of editing a node that doesn't
yet have a caption, any labels, or any properties. The previous approach
used the trick of modifying the underlying graph when starting to edit
these nodes, but that has some unfortunate snags. In particular, if the
user deleted an existing caption (for example by pressing backspace)
then the whole editor would disappear and they would not be able to type
in a replacement caption.

This new solution leaves the underlying graph unchanged, but instead
hacks at the VisualNode level. It doesn't modify the VisualNode, but
instead pretends that it would have created an additional caption
component, which can then be adapted by a CaptionEditor.