Stack components outside the node.
Previously caption, labels and properties were all positioned
independently outside the node, each component in turn trying to find
an orientation with the most space. Now the outside components use the
same model as the inside components: they stack on top of each other.

This change has some aesthetic advantages, making nodes in general look
much cleaner.

In later changes we need to clean up the style attributes so that there
is just a single attribute for the orientation of the outside components
as a whole instead of only having such a property for caption (and
this change make that property work a bit strangely).