Rename RelationshipCaption to RelationshipType.
The 'caption' name comes from a time before node labels were added to
the model, so essentially everything was a caption. Really this does
represent a relationship type, and it's more similar to a label than
to a caption in the node domain, so continuing to call it a caption is
confusing.

At some point we might decide to introduce something additional called
a relationship caption, but this will be in addition to relationship
type, not instead of it.

As a reminder, captions don't exist in the property graph model itself
but they are convenient way to identify nodes on top of the model. In
cypher-land, captions map to variable names, so if we ever introduce
relationship captions, we could use them in the same way.