Apply the component-stack model to relationships.
This is the same model used for nodes, now applied to relationships.
This means that orientation and position of type and properties now
appear directly underneath each other in all circumstances, and are
tied together with a single pair of style attributes. However, the
new style attributes don't yet have an effect - that needs to happen
in another commit.