Request fewer Google Drive permissions.
Request `drive.file` instead of `file`. The latter is a restricted
scope that would require this app to be verified and approved by
Google before widespread public use. Therefore, we only request
access to files created or opened by this app.

An unfortunate side effect of this change is that thumbnails don't
appear in the file picker. Therefore this commit falls back to a simple
list view for the picker, so that the the thumbnail problem is not so
obvious. I think the thumbnail problem is a bug on Google's side but
I have low confidence that we will be able to get them to fix it.