Hopeful quick fix for many-files problem.
A few people have reported a bug where they end up with a new file
for each individual change. I can't quite figure out how that happens,
but this change ought to help. In case the fileId doesn't get stored
in state at the first attempt, this gets it the next time. So in theory
there will only be two files at most.