Co-linear interval snapping.
If a dragged node is in the same row or column as other nodes, try to
achieve a consistent spacing of nodes within that row or column.

This works in two modes - either at the 'end' of sequence of equally
spaced nodes, or exactly half way between its two nearest neighbours.

This feature makes it easier to achieve grid-like layouts, without the
need to compare relationship lengths.