Include relationships in VisualGraph bounding box.
This was particularly a problem for self-relationships which always
extend beyond the bounding box of their node.

This behaviour is still not perfect - if the relationship has no
type or properties then only the mid point of the relationship is
considered, which results in a small amount visible relationship
being cropped.