Don't update graph if nodes have not moved.
Unfortunately, mouseUp on a single node is dispatching
MOVE_NODES_END_DRAG. This had the effect of marking the graph as 'dirty'
even if the user's only action was to click on a node to select it -
they hadn't really moved any nodes at all. This in turn led to lots
of unnecessary updates to Google Drive.

With this change we now recognise the case when nodes haven't really
been moved at all. It's not the ideal solution, but it seems to work.