Show all possible guides.
The maximum number of guides that constrain position is 2. However,
there may be many more guides that would have computed the same
position. Show all of these guides on screen (potentially hundreds of
them!) to make it clear that they exist, and to avoid having to explain
which two guides were the ones that happened to have been used to
compute position.