Avoid high pixel ratios for large diagrams.
Rendering PNGs for large diagrams is slow and blocks the main thread.
This change does fix the underlying problem, but it should reduce
render time by about 85% ((2+4)/(1+2+4) in the worst cases, by skipping
the 2x and 4x resolutions if the total number of pixels exceeds
16 million.

To fix the ultimate problem, we would need to move rendering onto a web
worker.